visualisation absences: ajout de listerSoldes + correction listerAbsences + gestion des exceptions

+++ src/main/java/dev/repository/CollegueRepo.java
@@ -6,7 +6,7 @@ import dev.entites.Collegue;
 
 import java.util.Optional;
 
-public interface CollegueRepo extends JpaRepository<Collegue, Long> {
+public interface CollegueRepo extends JpaRepository<Collegue, Integer> {
 
     Optional<Collegue> findByEmail(String email);
 }
